# internal/app/openapi.yaml
openapi: 3.1.0
info:
  title: Kore API
  version: 1.0.0
  description: API Kore — monolithe modulaire PSA/ESN
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /auth/login:
    post:
      summary: Authentification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password]
              properties:
                login: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Tokens émis
  /auth/refresh:
    post:
      summary: Renouvellement token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: Nouveaux tokens
...
  /societes:
    get:
      summary: Liste des sociétés
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Liste
    post:
      summary: Créer une société
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Créée
  /societes/{id}/branding:
    put:
      summary: Mettre à jour le branding tenant
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Branding mis à jour
  /users:
    get:
      summary: Liste utilisateurs
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Liste
    post:
      summary: Créer un utilisateur
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Créé
  /users/{id}:
    put:
      summary: Mettre à jour un utilisateur
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Mis à jour
    delete:
      summary: Supprimer un utilisateur (soft delete)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Supprimé
  /users/{id}/deactivate:
    patch:
      summary: Désactiver un utilisateur
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Désactivé
...
  /clients:
    get:
      summary: Liste clients
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Liste
    post:
      summary: Créer client
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Créé
...
  /sites:
    post:
      summary: Créer un site
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Site créé
  /services:
    post:
      summary: Créer un service
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Service créé
  /applications:
    post:
      summary: Créer une application
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Application créée
  /branding/logo/{tenantId}:
    get:
      summary: Logo du tenant
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Logo
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT